Remove pages around for stricter app definition
