+++ chateval/templates/registration/login.html
@@ -21,7 +21,7 @@
         {{ form.as_p }}
         <br>
         <button class="button" type="submit">Log In</button> <br> <br>
-        <a href="/accounts/signup"> Don't have an account? </a> 
+        <a href="/signup"> Don't have an account? </a> 
       </form>
     </div>
     <div class="column is-one-third"></div>
