+++ core/views.py
@@ -1,7 +1,6 @@
-from django.http import HttpResponseRedirect
 from django.shortcuts import render, redirect
-from .models import Baseline, EvaluationDataset, Model
-from .scripts import upload_model, load_responses
+from orm.models import Baseline, EvaluationDataset, Model
+from orm.scripts import get_messages
 
 def splash(request):
     datasets = EvaluationDataset.objects.all()
@@ -19,31 +18,17 @@ def conversations(request):
     return render(request, 'conversations.html', 
         {'POST': False, 'messages': messages, 'models': models, 'datasets': datasets})
 
-def submit(request):
-    response_files = EvaluationDataset.objects.all()
+def models(request):
+    models = Model.objects.all()
+    datasets = EvaluationDataset.objects.all()
+    messages = list()
+    evaluations = list()
     if request.method == "POST":
-        form = UploadModelForm(request.POST, request.FILES)
-        if form.is_valid():
-            model = Model(author=Author.objects.get(pk=request.user),
-                name=form.cleaned_data['name'],
-                description=form.cleaned_data['description'],
-                repo_location = form.cleaned_data['repo_location'],
-                cp_location=form.cleaned_data['checkpoint_location'])
-            model.save()
-            
-            files = list()
-            for response_file in response_files:
-                file = dict()
-                if request.FILES.get(response_file.name) is not None:
-                    file['file'] = request.FILES.get(response_file.name)
-                    file['dataset'] = response_file
-                    files.append(file)
-                    
-                    if 'baseline' in form.data.keys():
-                        baseline = Baseline(model=model, evaluationdataset=response_file)
-                        baseline.save()
-
-            upload_model(model, files, 'baseline' in form.data.keys())
-            return HttpResponseRedirect('/evaluation/')
-    form = UploadModelForm()
-    return render(request, 'submit.html', {'form': form, 'response_files': response_files})   
+        messages = get_messages(request.POST['model_id'], request.POST['evalset_id'])
+        dataset = EvaluationDataset.objects.get(pk=request.POST['evalset_id'])
+        for automatic_evaluation in AutomaticEvaluation.objects.filter(model=request.POST['model_id'], 
+            evaluationdataset=request.POST['evalset_id']):      
+            evaluations.append(dict({'name': automatic_evaluation.metric.name, 'value': automatic_evaluation.value}))
+        return render(request, 'models.html', {'POST': True, 'model': Model.objects.get(pk=request.POST['model_id']),
+            'messages': messages , 'models': models, 'datasets': datasets, 'dataset': dataset, 'evaluations': evaluations})
+    return render(request, 'models.html', {'POST': False, 'models': models, 'datasets': datasets})
